Phase B-3: kx-warrant lib.rs split (Rule 3, behavior-preserving move-only)
Third Phase B refactor per Rule 3. Pure move; zero behavior change.

Before: src/lib.rs 1254 lines (3 small enums with impls + WarrantField +
Host + FsScope + NetScope + ToolGrant + ModelRoute + ResourceCeiling +
WarrantSpec + Role + ToolRequirement + 2 typed errors + intersect/narrow
+ check_tool_requirement + warrant_ref_of + role_id_of + 26 inline tests).

After: thin src/lib.rs 100 lines (crate attrs + crate doc + mod decls +
pub use re-exports + #[cfg(test)] mod tests;) + 9 sibling modules:

- classes.rs (MoteClass + ExecutorClass + FsMode + impls)
- fields.rs (WarrantField + Host)
- scope.rs (FsScope + NetScope + impls)
- spec.rs (ToolGrant + ModelRoute + ResourceCeiling + WarrantSpec +
  Role + ToolRequirement)
- errors.rs (NarrowingError + ToolDenied)
- narrow.rs (intersect + narrow — load-bearing pure fns)
- check.rs (check_tool_requirement)
- refs.rs (warrant_ref_of + role_id_of)
- tests.rs (the 26 inline unit tests; #[cfg(test)] from lib.rs)

Re-exports preserve every previously-public path. No downstream crate's
use statements needed editing (cargo build --workspace clean).

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-warrant: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL)
- cargo test -p kx-warrant: 26 + 3 + 9 + 4 = 42 (IDENTICAL)
- cargo clippy -p kx-warrant --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean

Note: tests.rs imports were reduced (NdClass is used only via method
return types, not directly named — clippy's unused-import warning
surfaced the trim opportunity).

Next Phase B slice: kx-context-assembler (949 lines; borderline by line
count — will split if it benefits clarity).
